requirements.txt
termcolor
colorama
dnspython
